Unificado de package dto
no es necesario un nuevo archivo, solo hay que agregarlo dentro del principal
