+++ api/src/packages/dtos/Package.dto.ts
@@ -23,3 +23,29 @@ export class PackageDto {
   @IsNumber()
   showId: number;
 }
+
+export class PackageOnUpdateDto {
+  @IsOptional()
+  @IsString()
+  name: string;
+
+  @IsOptional()
+  @IsNumber()
+  quantPeople: number;
+
+  @IsOptional()
+  @IsNumber()
+  hotelId: number;
+
+  @IsOptional()
+  @IsNumber()
+  ticketId: number;
+
+  @IsOptional()
+  @IsNumber()
+  insuranceId: number;
+
+  @IsOptional()
+  @IsNumber()
+  showId: number;
+}
