--- api/src/packages/dtos/PackageOnUpdate.dto.ts
@@ -1,27 +1,0 @@
-import { IsNumber, IsOptional, IsString } from 'class-validator';
-
-export class PackageOnUpdateDto {
-  @IsOptional()
-  @IsString()
-  name: string;
-
-  @IsOptional()
-  @IsNumber()
-  quantPeople: number;
-
-  @IsOptional()
-  @IsNumber()
-  hotelId: number;
-
-  @IsOptional()
-  @IsNumber()
-  ticketId: number;
-
-  @IsOptional()
-  @IsNumber()
-  insuranceId: number;
-
-  @IsOptional()
-  @IsNumber()
-  showId: number;
-}
